+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## v0.2.2-alpha
+### BREAKING CHANGES
+- Users need to setup their config with script and change according to
+their needs
+
+### Changed
+    - fix: redundant code
+    - refactor!: use XDG Base Directory
+    
 ## v0.2.1-alpha
 ### Changed
   - Add github action for release
@@ -15,6 +24,6 @@ All notable changes to this project will be documented in this file.
   
 ## v0.1.0-alpha
 ### What's Changed
-  - Feat/laptop desktop by @Cyber-Syntax in https://github.com/Cyber-Syntax/fedora-setup/pull/1
+  - Feat/laptop desktop by @Cyber-Syntax in https://github.com/Cyber-Syntax/fedora-setup/pull/2
   - fix: wrong way to write grub file by @Cyber-Syntax in https://github.com/Cyber-Syntax/fedora-setup/pull/2
   - Initial setup script for Fedora